moved config to internal folder
